requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-secret-manager==2.16.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
